// shelter-frontend/src/app/api/model/createZbiorkaDto.ts
/**
 * Shelter
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateZbiorkaDto {
    nazwa?: string;
    cel?: string;
    kwotaDocelowa?: number;
    kwotaZebrana?: number;
    zwierzeta?: number[];
}